syling for password and username completed, will need to refactor the functions for accurate return
