+++ src/App.js
@@ -1,22 +1,28 @@
-import React from 'react';
+import { useContext } from 'react';
 import { BrowserRouter, Route, Routes } from 'react-router-dom/dist';
 import FormSet from './Components/SettingsForm/SettingsForm';
 import ToDo from './Components/ToDo/ToDo.js';
-import Footer from './Components/Footer/Footer.js';
 import HeaderUI from './Components/Header/Header.js';
-export default class App extends React.Component {
-  render() {
-    return (
-      <>
-        <BrowserRouter>
-          <HeaderUI />
+import { AuthContext } from './Context/Auth/Auth';
+import { When } from 'react-if';
+import Footer from './Components/Footer/Footer.js';
+const App = () => {
+  const { isLoggedIn } = useContext(AuthContext);
+
+  return (
+    <>
+      <BrowserRouter>
+        <HeaderUI />
+        <When condition={isLoggedIn}>
           <Routes>
-          <Route path="/"  element={<ToDo/>} />
-          <Route path="/settings"  element={<FormSet/>} />
-        </Routes>
+            <Route path="/" element={<ToDo />} />
+            <Route path="/settings" element={<FormSet />} />
+          </Routes>
+          </When>
           <Footer />
-        </BrowserRouter>
-      </>
-    );
-  }
+      </BrowserRouter>
+    </>
+  );
 }
+
+export default App;
